Set exclude_labels to True for kspixel pipeline
